move TitleBadge and BadgePosition definitions to docs aspect overview
