+++ scopes/component/ui/component-meta/index.ts
@@ -1,2 +1,2 @@
-export { ComponentOverview, BadgePosition } from './component-overview';
-export type { ComponentOverviewProps, TitleBadge } from './component-overview';
+export { ComponentOverview } from './component-overview';
+export type { ComponentOverviewProps } from './component-overview';
